Version bump to 7.0 release
Had to adjust the depends for this as a new QA notice on portage
revealed a missing dep. Linux trace tools are probably only required
when doing some tracing functions in PowerShell, but I couldn't find a
way to silence portage as it is configured as a FEATURE and those are
set in make.conf.